chess/src/main/java/ru/job4j/chess/firuges/black/BishopBlack.java
package ru.job4j.chess.firuges.black;

import ru.job4j.chess.ImpossibleMoveException;
import ru.job4j.chess.firuges.Cell;
import ru.job4j.chess.firuges.Figure;

public class BishopBlack implements Figure {
    private final Cell position;

    public BishopBlack(final Cell ps) {
        position = ps;
    }

    @Override
    public Cell position() {
        return position;
    }

    @Override
    public Cell[] way(Cell source,Cell dest) {
        if (!isDiagonal(source, dest)) {
            throw new ImpossibleMoveException(
                    String.format("Could not move by diagonal from %s to %s", source, dest)
            );
        }
        int size = Math.abs(source.getX() - dest.getX());
        Cell[] steps = new Cell[size];
        int deltaX = dest.getX() - source.getX() > 0 ? 1 : -1;
        int deltaY = dest.getY() - source.getY() > 0 ? 1 : -1;
        for (int index = 0; index < size; index++) {
            int x = source.getX() + (index + 1) * deltaX;;
            int y = source.getY() + (index + 1) * deltaY;;
            steps[index] = Cell.findBy(x,y);
        }
        return steps;


    }

    public boolean isDiagonal(Cell source, Cell dest) {
        /* TODO check diagonal */
          int size =  Math.abs(source.getX() - dest.getX());
          if (Math.abs(source.getX() - dest.getX())==size&&Math.abs(source.getY() - dest.getY())==size){
              return true;
          }
        return false;
    }

    @Override
    public Figure copy(Cell dest) {
        return new BishopBlack(dest);
    }
}
